[Feature] : timestampToString 구현 및 테스트

+++ ExchangeRateCalcApp/ExchangeRateCalcApp/Domain/Entity/ExchangeRateInformationDTO.swift
@@ -11,7 +11,7 @@ struct ExchangeRateInformationDTO: Decodable {
     let success: Bool?
     let terms: String?
     let privacy: String?
-    let timestamp: Int?
+    let timestamp: Double?
     let source: String?
     let quotes: Quotes
 }
